requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap-form==3.4
django-bootstrap4==2.3.1
django-filter==2.4.0
django-oauth-toolkit==1.4.0
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
